refactor(interactions): migrate trial buttons + modals to registry
Four duplicate officer-role-check blocks collapse to officerOnly:true on
each trial button handler entry.
